finish recursion

+++ src/LineSeg.cpp
@@ -87,18 +87,18 @@ Cross_Param(LineSeg e)
 
 // Returns 1 if the lines intersect, otherwise 0. In addition, if the lines 
 // intersect the intersection point may be stored in the floats i_x and i_y.
-bool LineSeg::get_line_intersection(float p0_x, float p0_y, float p1_x, float p1_y, 
-    float p2_x, float p2_y, float p3_x, float p3_y, float *i_x, float *i_y)
+bool LineSeg::get_line_intersection(double p0_x, double p0_y, double p1_x, double p1_y, 
+    double p2_x, double p2_y, double p3_x, double p3_y, double *i_x, double *i_y)
 {
-    float s1_x, s1_y, s2_x, s2_y;
+    double s1_x, s1_y, s2_x, s2_y;
     s1_x = p1_x - p0_x;     s1_y = p1_y - p0_y;
     s2_x = p3_x - p2_x;     s2_y = p3_y - p2_y;
 
-    float s, t;
+    double s, t;
     s = (-s1_y * (p0_x - p2_x) + s1_x * (p0_y - p2_y)) / (-s2_x * s1_y + s1_x * s2_y);
     t = ( s2_x * (p0_y - p2_y) - s2_y * (p0_x - p2_x)) / (-s2_x * s1_y + s1_x * s2_y);
 
-    if (s >= 0 && s <= 1 && t >= 0 && t <= 1)
+    if (s >= 0.0 && s <= 1.0 && t >= 0.0 && t <= 1.0)
     {
         // Collision detected
         if (i_x != 0)
