+++ src/LineSeg.h
@@ -39,12 +39,12 @@ class LineSeg {
 		// DOES NOT tell you whether the edge is "entering" or "leaving".
 		// But you can use tests like Edge::Point_Side() to figure that out.
 		float   Cross_Param(LineSeg);
-		static bool get_line_intersection(float p0_x, float p0_y, float p1_x, float p1_y, 
-     	  float p2_x, float p2_y, float p3_x, float p3_y, float *i_x, float *i_y);
+		static bool get_line_intersection(double p0_x, double p0_y, double p1_x, double p1_y, 
+     	  double p2_x, double p2_y, double p3_x, double p3_y, double *i_x, double *i_y);
 
 	public:
-		float   start[2];	// Starting point, x and y.
-		float   end[2];	// Ending point, x and y.
+		double   start[2];	// Starting point, x and y.
+		double   end[2];	// Ending point, x and y.
 };
 
 #endif
